requirements.txt
pyTelegramBotAPI
openai
psycopg2-binary
stripe
flask
gunicorn